refactor(runner): give the dependency cache entry store its own interface
The content-addressed store on disk -- entry layout, publication, immutability
validation, lock, usage markers, byte accounting, eviction and rollback -- was
about 400 lines inside dependency-cache.ts, reachable only through
materializeWorkspaceDependencies. Every retention test therefore had to build
an npm workspace and drive a fake install executor to get an entry on disk.

dependency-cache-store.ts now owns it, with keys, entries, bytes and usage as
its vocabulary and no RunnerConfig, env or executor in its interface. Entry
validation applies its confinement rule against a nominal restore root, which
is lexically the same answer as the run workspace gave it, so the store needs
no workspace at all. dependency-cache.ts keeps key derivation, npm execution
and orchestration, and reaches the store through a handle bound to one resolved
cache root.

The eight retention tests that only needed entries on disk move to
dependency-cache-store.test.ts and run on a temporary directory; the module
cast that reached for the old file's unexported retention functions is gone.
What stays in dependency-cache.test.ts is the interaction that needs both:
enforcement waiting behind a materialization's shared lock.
