+++ packages/runner/src/dependency-cache.ts
@@ -1,56 +1,20 @@
 import { createHash } from "node:crypto";
-import { constants } from "node:fs";
-import {
-  chmod, cp, lstat, mkdir, mkdtemp, open, readFile, readdir, readlink, realpath, rename, rm, writeFile,
-} from "node:fs/promises";
+import { lstat, readFile, readdir, realpath } from "node:fs/promises";
 import { platform } from "node:os";
 import { dirname, isAbsolute, join, posix, relative, resolve, sep } from "node:path";
 
-import { flock } from "fs-ext";
-
 import type { RunnerConfig } from "./config.js";
+import {
+  CACHE_ENTRY_FORMAT, DependencyCacheIntegrityError, assertCacheTargetPath, describeTargetTrees, openCacheEntryStore,
+  type CacheEntryDocument, type CacheEntryExpectation, type CacheEntryInput, type CacheEntryStore,
+  type CacheEntryTarget, type DependencyCacheToolchain,
+} from "./dependency-cache-store.js";
 import type { CommandOptions } from "./exec.js";
 import {
   NPM_INSTALL_COMMAND_TIMEOUT_MS, NPM_INSTALL_OPERATION_BUDGET_MS, runWithNetworkRetry, type RetryOptions,
 } from "./network-retry.js";
 
-const CACHE_FORMAT = "agentos-runner-dependency-cache-v2";
-const METADATA_FILE = "metadata.json";
-const TREE_DIRECTORY = "trees";
-const MAX_METADATA_BYTES = 128 * 1024 * 1024;
 const NPM_PROBE_TIMEOUT_MS = 10_000;
-const CACHE_LOCK_FILE = "lock";
-const CACHE_USAGE_DIRECTORY = "usage";
-const CACHE_KEY = /^[a-f0-9]{64}$/u;
-export const DEPENDENCY_CACHE_BYTE_BUDGET = 16 * 1024 ** 3;
-
-export type DependencyCacheRetentionSize = { key: string; bytes: number; usedMs: number };
-
-export const selectDependencyCacheEvictions = (
-  entries: readonly DependencyCacheRetentionSize[],
-  currentKey?: string,
-  budget = DEPENDENCY_CACHE_BYTE_BUDGET,
-): string[] => {
-  if (!Number.isFinite(budget) || budget < 0) throw new Error("Dependency cache byte budget is invalid");
-  for (const entry of entries) {
-    if (!Number.isFinite(entry.bytes) || entry.bytes < 0) throw new Error("Dependency cache entry size is invalid");
-    if (!Number.isFinite(entry.usedMs)) throw new Error("Dependency cache usage time is invalid");
-  }
-  const protectedEntry = currentKey === undefined ? undefined : entries.find(({ key }) => key === currentKey);
-  if (protectedEntry !== undefined && protectedEntry.bytes > budget) {
-    throw new DependencyCacheBudgetError("protected-entry-exceeds-byte-budget");
-  }
-  let total = entries.reduce((sum, entry) => sum + entry.bytes, 0);
-  const victims: string[] = [];
-  const ordered = [...entries].sort((left, right) => left.usedMs - right.usedMs || left.key.localeCompare(right.key));
-  for (const entry of ordered) {
-    if (total <= budget) break;
-    if (entry.key === currentKey) continue;
-    total -= entry.bytes;
-    victims.push(entry.key);
-  }
-  return victims;
-};
 
 export type DependencyCommandExecutor = (
   config: RunnerConfig,
@@ -65,30 +29,8 @@ export type DependencyCacheDependencies = {
   execute: DependencyCommandExecutor;
 };
 
-export type DependencyCacheToolchain = {
-  node: string;
-  npm: string;
-  operatingSystem: string;
-  architecture: string;
-};
-
-type CacheInput = { path: string; sha256: string } | { path: string; absent: true };
-type TreeManifestEntry =
-  | { path: string; kind: "directory"; mode: number }
-  | { path: string; kind: "file"; mode: number; sha256: string }
-  | { path: string; kind: "symlink"; target: string };
-type TargetManifestEntry = { path: string; present: boolean; tree: TreeManifestEntry[] };
-
-type CacheMetadata = {
-  format: typeof CACHE_FORMAT;
-  key: string;
-  toolchain: DependencyCacheToolchain;
-  inputs: CacheInput[];
-  targets: TargetManifestEntry[];
-};
-
 type DependencyProject = {
-  inputs: CacheInput[];
+  inputs: CacheEntryInput[];
   targets: string[];
   nativeWorkspaces: string[];
   inputMissCondition?: string;
@@ -149,20 +91,6 @@ class DependencyInputMissSignal extends Error {
   }
 }
 
-export class DependencyCacheIntegrityError extends Error {
-  constructor(readonly condition: string) {
-    super(`Dependency cache integrity refusal: ${condition}`);
-    this.name = "DependencyCacheIntegrityError";
-  }
-}
-
-export class DependencyCacheBudgetError extends Error {
-  constructor(readonly condition: string) {
-    super(`Dependency cache byte budget refusal: ${condition}`);
-    this.name = "DependencyCacheBudgetError";
-  }
-}
-
 const insideOrEqual = (root: string, candidate: string): boolean =>
   candidate === root || candidate.startsWith(`${root}${sep}`);
 
@@ -200,7 +128,7 @@ const pathKind = async (path: string): Promise<"missing" | "directory" | "file" 
   }
 };
 
-const readInput = async (workspace: string, path: string, required: boolean): Promise<CacheInput> => {
+const readInput = async (workspace: string, path: string, required: boolean): Promise<CacheEntryInput> => {
   const absolute = resolve(workspace, path);
   if (!insideOrEqual(workspace, absolute)) throw new Error(`Ambiguous dependency input: ${path}`);
   const kind = await pathKind(absolute);
@@ -430,7 +358,7 @@ const inspectDependencyProject = async (workspacePath: string): Promise<Dependen
     if (!pkg.name) throw new Error(`Native workspace ${pkg.manifestPath} has no package name`);
     nativeWorkspaces.push(pkg.name);
   }
-  const inputs: CacheInput[] = [];
+  const inputs: CacheEntryInput[] = [];
   let inputMissCondition: string | undefined;
   for (const { path, required } of [
     ...requiredPaths.map((path) => ({ path, required: true })),
@@ -521,7 +449,7 @@ const effectiveNpmConfigInput = async (
   workspace: string,
   env: NodeJS.ProcessEnv,
   execute: DependencyCommandExecutor,
-): Promise<CacheInput> => {
+): Promise<CacheEntryInput> => {
   const raw = await execute(config, "npm", ["config", "ls", "--json"], workspace, env, { timeoutMs: NPM_PROBE_TIMEOUT_MS });
   const parsed = parseJsonObject(raw, "effective npm configuration");
   const filtered = canonicalValue(parsed);
@@ -531,7 +459,7 @@ const effectiveNpmConfigInput = async (
 export type DependencyCacheIdentity = {
   key: string;
   toolchain: DependencyCacheToolchain;
-  inputs: CacheInput[];
+  inputs: CacheEntryInput[];
   targets: string[];
   nativeWorkspaces: string[];
 };
@@ -556,17 +484,8 @@ export const deriveDependencyCacheKey = async (
   inputs.push(await effectiveNpmConfigInput(config, workspace, env, execute));
   inputs.sort((left, right) => left.path.localeCompare(right.path, "en"));
   const toolchain = options.toolchain ?? await currentToolchain(config, workspace, env, execute);
-  const key = sha256(`${JSON.stringify({ format: CACHE_FORMAT, toolchain: validatedToolchain(toolchain), inputs })}\n`);
+  const key = sha256(`${JSON.stringify({ format: CACHE_ENTRY_FORMAT, toolchain: validatedToolchain(toolchain), inputs })}\n`);
   return { key, toolchain, inputs, targets, nativeWorkspaces };
-};
-
-const assertTarget = (workspace: string, target: string): string => {
-  if (target !== "node_modules" && !target.endsWith("/node_modules")) {
-    throw new Error(`Invalid dependency target: ${target}`);
-  }
-  const absolute = resolve(workspace, target);
-  if (!insideOrEqual(workspace, absolute)) throw new Error(`Dependency target escaped the run workspace: ${target}`);
-  return absolute;
 };
 
 const clearTargets = async (
@@ -578,7 +497,7 @@ const clearTargets = async (
 ): Promise<void> => {
   const existing: Array<{ absolute: string; target: string }> = [];
   for (const target of targets) {
-    const absolute = assertTarget(workspace, target);
+    const absolute = assertCacheTargetPath(workspace, target);
     await ensureRealDirectory(workspace, posix.dirname(target), `Dependency target parent ${posix.dirname(target)}`);
     const kind = await pathKind(absolute);
     if (kind === "symlink") throw new Error(`Dependency target is a symlink: ${target}`);
@@ -591,689 +510,44 @@ const clearTargets = async (
   }
 };
 
-const treeManifest = async (
-  treeRoot: string,
-  workspace: string,
-  workspaceTarget: string,
-  requireImmutable: boolean,
-): Promise<TreeManifestEntry[]> => {
-  const manifest: TreeManifestEntry[] = [];
-  const visit = async (path: string): Promise<void> => {
-    const info = await lstat(path);
-    const mapped = resolve(workspaceTarget, relative(treeRoot, path));
-    if (!insideOrEqual(workspace, mapped)) throw new DependencyCacheIntegrityError("tree-path-escape");
-    const manifestPath = relative(treeRoot, path).split(sep).join("/") || ".";
-    if (info.isSymbolicLink()) {
-      const link = await readlink(path);
-      if (isAbsolute(link) || !insideOrEqual(workspace, resolve(dirname(mapped), link))) {
-        throw new DependencyCacheIntegrityError("symlink-escape");
-      }
-      manifest.push({ path: manifestPath, kind: "symlink", target: link });
-      return;
-    }
-    if (requireImmutable && (info.mode & 0o222) !== 0) {
-      throw new DependencyCacheIntegrityError("writable-entry");
-    }
-    if (info.isDirectory()) {
-      if (requireImmutable && (info.mode & 0o005) !== 0o005) {
-        throw new DependencyCacheIntegrityError("entry-not-readable");
-      }
-      manifest.push({ path: manifestPath, kind: "directory", mode: info.mode & 0o111 });
-      for (const child of (await readdir(path)).sort()) await visit(join(path, child));
-      return;
-    }
-    if (!info.isFile()) throw new DependencyCacheIntegrityError("special-file");
-    if (requireImmutable && (info.mode & 0o004) === 0) throw new DependencyCacheIntegrityError("entry-not-readable");
-    manifest.push({ path: manifestPath, kind: "file", mode: info.mode & 0o111, sha256: sha256(await readFile(path)) });
-  };
-  await visit(treeRoot);
-  return manifest;
-};
+const sameJson = (left: unknown, right: unknown): boolean => JSON.stringify(left) === JSON.stringify(right);
 
-const findTargetManifest = async (workspace: string, targets: string[]): Promise<TargetManifestEntry[]> => {
-  const manifest: TargetManifestEntry[] = [];
-  for (const target of targets) {
-    const absolute = assertTarget(workspace, target);
-    const kind = await pathKind(absolute);
-    if (kind === "symlink") throw new Error(`Dependency target is a symlink: ${target}`);
-    if (kind !== "missing" && kind !== "directory") throw new Error(`Dependency target is not a directory: ${target}`);
-    const tree = kind === "directory" ? await treeManifest(absolute, workspace, absolute, false) : [];
-    manifest.push({ path: target, present: kind === "directory", tree });
-  }
+// npm ci is the only producer of these trees, so the root target must exist
+// afterwards. An install that leaves it missing is a failed install, not an
+// entry worth publishing.
+const installedTargetTrees = async (workspace: string, targets: string[]): Promise<CacheEntryTarget[]> => {
+  const manifest = await describeTargetTrees(workspace, targets);
   if (!manifest[0]?.present || manifest[0].path !== "node_modules") {
     throw new Error("npm ci did not produce the root node_modules target");
   }
   return manifest;
 };
 
-const metadataShape = (value: unknown): value is CacheMetadata => {
-  if (value === null || typeof value !== "object" || Array.isArray(value)) return false;
-  const metadata = value as Partial<CacheMetadata>;
-  if (metadata.format !== CACHE_FORMAT || typeof metadata.key !== "string") return false;
-  if (metadata.toolchain === null || typeof metadata.toolchain !== "object" || Array.isArray(metadata.toolchain)) return false;
-  const toolchain = metadata.toolchain as Partial<DependencyCacheToolchain>;
-  if (!sameJson(Object.keys(toolchain).sort(), ["architecture", "node", "npm", "operatingSystem"])) return false;
-  if (![toolchain.node, toolchain.npm, toolchain.operatingSystem, toolchain.architecture]
-    .every((coordinate) => typeof coordinate === "string" && coordinate.length > 0)) return false;
-  if (!Array.isArray(metadata.inputs) || !Array.isArray(metadata.targets) || metadata.targets.length === 0) return false;
-  return metadata.inputs.every((input) => {
-    if (input === null || typeof input !== "object" || typeof (input as CacheInput).path !== "string") return false;
-    const fields = Object.keys(input).sort();
-    return (sameJson(fields, ["path", "sha256"]) && /^[a-f0-9]{64}$/u.test(String((input as { sha256?: unknown }).sha256)))
-      || (sameJson(fields, ["absent", "path"]) && (input as { absent?: unknown }).absent === true);
-  })
-    && metadata.targets.every((target) => target !== null && typeof target === "object"
-      && typeof (target as TargetManifestEntry).path === "string"
-      && typeof (target as TargetManifestEntry).present === "boolean"
-      && Array.isArray((target as TargetManifestEntry).tree)
-      && sameJson(Object.keys(target).sort(), ["path", "present", "tree"])
-      && (target as TargetManifestEntry).tree.every((entry) => {
-        if (entry === null || typeof entry !== "object" || typeof entry.path !== "string" || typeof entry.kind !== "string") return false;
-        if (entry.kind === "directory") {
-          return Number.isInteger(entry.mode) && entry.mode >= 0 && entry.mode <= 0o111
-            && sameJson(Object.keys(entry).sort(), ["kind", "mode", "path"]);
-        }
-        if (entry.kind === "file") {
-          return Number.isInteger(entry.mode) && entry.mode >= 0 && entry.mode <= 0o111
-            && /^[a-f0-9]{64}$/u.test(entry.sha256)
-            && sameJson(Object.keys(entry).sort(), ["kind", "mode", "path", "sha256"]);
-        }
-        return entry.kind === "symlink" && typeof entry.target === "string"
-          && sameJson(Object.keys(entry).sort(), ["kind", "path", "target"]);
-      }));
-};
-
-const readMetadata = async (entry: string): Promise<CacheMetadata> => {
-  const path = join(entry, METADATA_FILE);
-  const info = await lstat(path).catch((error: unknown) => {
-    if (errorCode(error) === "ENOENT") throw new DependencyCacheIntegrityError("metadata-missing");
-    throw error;
-  });
-  if (!info.isFile() || info.isSymbolicLink() || info.size > MAX_METADATA_BYTES) {
-    throw new DependencyCacheIntegrityError("metadata-malformed");
-  }
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(await readFile(path, "utf8"));
-  } catch (error: unknown) {
-    throw new DependencyCacheIntegrityError(`metadata-unreadable:${errorCode(error) ?? "invalid-json"}`);
-  }
-  if (!metadataShape(parsed)) throw new DependencyCacheIntegrityError("metadata-malformed");
-  return parsed;
-};
-
-const sameJson = (left: unknown, right: unknown): boolean => JSON.stringify(left) === JSON.stringify(right);
-
-const validateTreeLayout = async (trees: string, targets: TargetManifestEntry[]): Promise<void> => {
-  const present = new Set(targets.filter(({ present }) => present).map(({ path }) => path));
-  const prefixes = new Set<string>();
-  for (const target of present) {
-    let parent = posix.dirname(target);
-    while (parent !== ".") {
-      prefixes.add(parent);
-      parent = posix.dirname(parent);
-    }
-  }
-  const walk = async (directory: string, relativeDirectory = ""): Promise<void> => {
-    if (((await lstat(directory)).mode & 0o222) !== 0) throw new DependencyCacheIntegrityError("writable-entry");
-    for (const entry of await readdir(directory, { withFileTypes: true })) {
-      const path = relativeDirectory === "" ? entry.name : `${relativeDirectory}/${entry.name}`;
-      if (present.has(path)) continue;
-      if (!prefixes.has(path) || !entry.isDirectory() || entry.isSymbolicLink()) {
-        throw new DependencyCacheIntegrityError("unexpected-tree-content");
-      }
-      await walk(join(directory, entry.name), path);
-    }
-  };
-  await walk(trees);
-};
-
-const validateImmutableEntryContents = async (
-  entry: string,
-  metadata: CacheMetadata,
-  workspace: string,
-): Promise<void> => {
-  const trees = join(entry, TREE_DIRECTORY);
-  if (await pathKind(trees) !== "directory") throw new DependencyCacheIntegrityError("tree-root-missing");
-  const entryNames = (await readdir(entry)).sort();
-  if (!sameJson(entryNames, [METADATA_FILE, TREE_DIRECTORY].sort())) {
-    throw new DependencyCacheIntegrityError("unexpected-entry-content");
-  }
-  await validateTreeLayout(trees, metadata.targets);
-  for (const target of metadata.targets) {
-    const cached = resolve(trees, target.path);
-    if (!insideOrEqual(trees, cached)) throw new DependencyCacheIntegrityError("target-path-escape");
-    const kind = await pathKind(cached);
-    if (target.present && kind !== "directory") throw new DependencyCacheIntegrityError("target-tree-missing");
-    if (!target.present && kind !== "missing") throw new DependencyCacheIntegrityError("unexpected-target-tree");
-    if (target.present) {
-      const actualTree = await treeManifest(cached, workspace, assertTarget(workspace, target.path), true);
-      if (!sameJson(actualTree, target.tree)) throw new DependencyCacheIntegrityError("tree-manifest-mismatch");
-    } else if (target.tree.length !== 0) {
-      throw new DependencyCacheIntegrityError("target-manifest-mismatch");
-    }
-  }
-  await treeManifest(join(entry, METADATA_FILE), workspace, workspace, true);
-  const entryInfo = await lstat(entry);
-  const treesInfo = await lstat(trees);
-  if ((entryInfo.mode & 0o222) !== 0 || (treesInfo.mode & 0o222) !== 0) {
-    throw new DependencyCacheIntegrityError("writable-entry");
-  }
-};
-
-const validateEntry = async (
-  entry: string,
-  expected: Omit<CacheMetadata, "targets"> & { targetPaths: string[] },
-  workspace: string,
-): Promise<CacheMetadata> => {
-  if (await pathKind(entry) !== "directory") throw new DependencyCacheIntegrityError("entry-not-directory");
-  const metadata = await readMetadata(entry);
-  if (metadata.key !== expected.key) throw new DependencyCacheIntegrityError("key-mismatch");
-  if (!sameJson(metadata.toolchain, expected.toolchain)) throw new DependencyCacheIntegrityError("toolchain-mismatch");
-  if (!sameJson(metadata.inputs, expected.inputs)) throw new DependencyCacheIntegrityError("input-manifest-mismatch");
-  if (!sameJson(metadata.targets.map(({ path }) => path), expected.targetPaths)) {
-    throw new DependencyCacheIntegrityError("target-manifest-mismatch");
-  }
-  await validateImmutableEntryContents(entry, metadata, workspace);
-  return metadata;
-};
-
-const makeImmutable = async (path: string): Promise<void> => {
-  const info = await lstat(path);
-  if (info.isSymbolicLink()) return;
-  if (info.isDirectory()) for (const child of await readdir(path)) await makeImmutable(join(path, child));
-  await chmod(path, info.isDirectory() ? 0o555 : 0o444 | (info.mode & 0o111));
-};
-
-const makeWritable = async (path: string): Promise<void> => {
-  const kind = await pathKind(path);
-  if (kind === "missing" || kind === "symlink") return;
-  const info = await lstat(path);
-  await chmod(path, info.mode | 0o700);
-  if (kind === "directory") for (const child of await readdir(path)) await makeWritable(join(path, child));
-};
-
-const allocatedBytes = (info: { blocks: number }): bigint => {
-  if (!Number.isSafeInteger(info.blocks) || info.blocks < 0) {
-    throw new DependencyCacheIntegrityError("invalid-allocated-size");
-  }
-  return BigInt(info.blocks) * 512n;
-};
-
-// Cache entries are immutable after publication. Walk them without following
-// symlinks so accounting cannot escape the canonical cache root. A symlink is
-// itself an inode and therefore contributes its own allocated blocks. Its
-// lexical target must remain below the entry root; validateEntry additionally
-// checks the target against the workspace layout before a selected entry is
-// restored.
-const accountCacheEntry = async (entry: string): Promise<bigint> => {
-  const root = resolve(entry);
-  const rootInfo = await lstat(root).catch((error: unknown) => {
-    if (errorCode(error) === "ENOENT") throw new DependencyCacheIntegrityError("entry-not-directory");
-    throw error;
-  });
-  if (rootInfo.isSymbolicLink() || !rootInfo.isDirectory()) {
-    throw new DependencyCacheIntegrityError(rootInfo.isSymbolicLink() ? "unsafe-retention-entry" : "entry-not-directory");
-  }
-  const visitedInodes = new Set<string>();
-  const visit = async (path: string): Promise<bigint> => {
-    let info;
-    try {
-      info = await lstat(path);
-    } catch (error: unknown) {
-      if (errorCode(error) === "ENOENT") throw new DependencyCacheIntegrityError("entry-disappeared");
-      throw error;
-    }
-    const inode = `${String(info.dev)}:${String(info.ino)}`;
-    if (info.ino !== 0 && visitedInodes.has(inode)) return 0n;
-    if (info.ino !== 0) visitedInodes.add(inode);
-    const ownBytes = allocatedBytes(info);
-    if (info.isSymbolicLink()) {
-      const target = await readlink(path);
-      if (isAbsolute(target) || !insideOrEqual(root, resolve(dirname(path), target))) {
-        throw new DependencyCacheIntegrityError("symlink-escape");
-      }
-      return ownBytes;
-    }
-    if (info.isFile()) return ownBytes;
-    if (!info.isDirectory()) throw new DependencyCacheIntegrityError("special-file");
-    let total = ownBytes;
-    for (const child of await readdir(path)) {
-      const childPath = join(path, child);
-      if (!insideOrEqual(root, childPath)) throw new DependencyCacheIntegrityError("entry-path-escape");
-      total += await visit(childPath);
-    }
-    return total;
-  };
-  return visit(root);
-};
-
-export const accountDependencyCacheEntryBytes = async (entry: string): Promise<bigint> => {
-  let bytes: bigint;
-  try {
-    bytes = await accountCacheEntry(entry);
-  } catch (error: unknown) {
-    if (error instanceof DependencyCacheIntegrityError) throw error;
-    throw new DependencyCacheIntegrityError(`size-walk-failed:${errorCode(error) ?? "unknown"}`);
-  }
-  return bytes;
-};
-
-const ensureCacheRoot = async (configuredRoot: string, workspace: string): Promise<string> => {
-  const requestedRoot = resolve(configuredRoot);
-  await mkdir(requestedRoot, { recursive: true, mode: 0o711 });
-  if ((await lstat(requestedRoot)).isSymbolicLink()) throw new Error("Dependency cache root is a symlink");
-  const root = await realpath(requestedRoot);
-  if (insideOrEqual(root, workspace) || insideOrEqual(workspace, root)) {
-    throw new Error("Dependency cache root overlaps the run workspace");
-  }
-  await chmod(root, 0o711);
-  const entries = join(root, "entries");
-  await mkdir(entries, { recursive: true, mode: 0o711 });
-  if ((await lstat(entries)).isSymbolicLink()) throw new Error("Dependency cache entries root is a symlink");
-  await chmod(entries, 0o711);
-  const usage = join(root, CACHE_USAGE_DIRECTORY);
-  await mkdir(usage, { recursive: true, mode: 0o700 });
-  if ((await lstat(usage)).isSymbolicLink()) throw new Error("Dependency cache usage root is a symlink");
-  await chmod(usage, 0o700);
-  return root;
-};
-
-const flockAsync = (fd: number, operation: "sh" | "ex" | "un"): Promise<void> => new Promise((accept, reject) => {
-  flock(fd, operation, (error) => error ? reject(error) : accept());
-});
-
-const openCacheLock = async (cacheRoot: string) => {
-  const path = join(cacheRoot, CACHE_LOCK_FILE);
-  const handle = await open(path, constants.O_CREAT | constants.O_RDWR | constants.O_NOFOLLOW, 0o600);
-  const info = await handle.stat();
-  if (!info.isFile()) {
-    await handle.close();
-    throw new Error("Dependency cache lock is not a file");
-  }
-  return handle;
-};
-
-const withSharedCacheLock = async <T>(cacheRoot: string, work: () => Promise<T>): Promise<T> => {
-  const handle = await openCacheLock(cacheRoot);
-  try {
-    await flockAsync(handle.fd, "sh");
-    try {
-      return await work();
-    } finally {
-      await flockAsync(handle.fd, "un");
-    }
-  } finally {
-    await handle.close();
-  }
-};
-
-const recordCacheUse = async (cacheRoot: string, key: string): Promise<void> => {
-  if (!CACHE_KEY.test(key)) throw new Error("Dependency cache usage key is invalid");
-  const marker = join(cacheRoot, CACHE_USAGE_DIRECTORY, key);
-  const handle = await open(
-    marker,
-    constants.O_CREAT | constants.O_WRONLY | constants.O_TRUNC | constants.O_NOFOLLOW,
-    0o600,
-  ).catch((error: unknown) => {
-    throw new DependencyCacheIntegrityError(`usage-marker-unwritable:${errorCode(error) ?? "unknown"}`);
-  });
-  try {
-    await handle.writeFile(`${new Date().toISOString()}\n`);
-  } finally {
-    await handle.close();
-  }
-};
-
-const validateCacheUseMarker = async (cacheRoot: string, key: string): Promise<void> => {
-  const marker = join(cacheRoot, CACHE_USAGE_DIRECTORY, key);
-  const info = await lstat(marker).catch((error: unknown) => {
-    if (errorCode(error) === "ENOENT") return null;
-    throw new DependencyCacheIntegrityError(`usage-marker-unreadable:${errorCode(error) ?? "unknown"}`);
-  });
-  if (info?.isSymbolicLink() || (info !== null && !info.isFile())) {
-    throw new DependencyCacheIntegrityError("unsafe-usage-marker");
-  }
-};
-
-type RetentionEntry = { key: string; path: string; usedMs: number; bytes: bigint };
-
-const reportIntegrityRefusal = (
-  report: (progress: DependencyCacheProgress) => void,
-  key: string | undefined,
-  condition: string,
-): DependencyCacheIntegrityError => {
-  report({ event: "integrity-refusal", ...(key ? { key: key.slice(0, 16) } : {}), condition });
-  return new DependencyCacheIntegrityError(condition);
-};
-
-const inspectRetentionEntry = async (key: string, path: string): Promise<void> => {
-  const metadata = await readMetadata(path);
-  if (metadata.key !== key) throw new DependencyCacheIntegrityError("key-mismatch");
-  // Retention has no originating workspace, but the immutable manifest's
-  // relative targets can be mapped beneath a lexical workspace boundary. This
-  // applies the same target and symlink confinement as a selected restore.
-  await validateImmutableEntryContents(path, metadata, join(path, ".retention-workspace"));
-};
-
-const retentionEntries = async (
-  cacheRoot: string,
-  report: (progress: DependencyCacheProgress) => void,
-): Promise<RetentionEntry[]> => {
-  const entriesRoot = join(cacheRoot, "entries");
-  const usageRoot = join(cacheRoot, CACHE_USAGE_DIRECTORY);
-  if (await pathKind(entriesRoot) !== "directory") {
-    throw reportIntegrityRefusal(report, undefined, "entries-root-missing");
-  }
-  if (await pathKind(usageRoot) !== "directory") {
-    throw reportIntegrityRefusal(report, undefined, "usage-root-missing");
-  }
-  const markerKinds = new Map<string, Awaited<ReturnType<typeof lstat>> | null>();
-  for (const name of await readdir(usageRoot)) {
-    // Only key-shaped names are cache usage markers. Ignore unrelated files
-    // rather than letting them deny all materializations on the shared host.
-    if (!CACHE_KEY.test(name)) continue;
-    const marker = join(usageRoot, name);
-    const markerInfo = await lstat(marker).catch((error: unknown) => {
-      if (errorCode(error) === "ENOENT") return null;
-      throw error;
-    });
-    if (markerInfo?.isSymbolicLink() || (markerInfo !== null && !markerInfo.isFile())) {
-      throw reportIntegrityRefusal(report, name, "unsafe-usage-marker");
-    }
-    markerKinds.set(name, markerInfo);
-  }
-  const entries: RetentionEntry[] = [];
-  for (const name of await readdir(entriesRoot)) {
-    if (!CACHE_KEY.test(name)) {
-      // publishEntry creates stages under this directory while holding a
-      // shared lock. Once the exclusive retention lock is held, a remaining
-      // stage is necessarily orphaned and may be reaped safely. Other non-key
-      // names are not immutable cache entries and are ignored.
-      if (name.startsWith(".stage-")) {
-        const staging = join(entriesRoot, name);
-        await makeWritable(staging).catch(() => undefined);
-        await rm(staging, { recursive: true, force: true }).catch(() => undefined);
-      }
-      continue;
-    }
-    const path = join(entriesRoot, name);
-    const info = await lstat(path).catch((error: unknown) => {
-      if (errorCode(error) === "ENOENT") return null;
-      throw error;
-    });
-    if (info === null) throw reportIntegrityRefusal(report, name, "retention-entry-disappeared");
-    if (info.isSymbolicLink() || !info.isDirectory()) {
-      throw reportIntegrityRefusal(report, name, "unsafe-retention-entry");
-    }
-    let bytes: bigint;
-    try {
-      await inspectRetentionEntry(name, path);
-      bytes = await accountCacheEntry(path);
-    } catch (error: unknown) {
-      const integrityError = error instanceof DependencyCacheIntegrityError
-        ? error
-        : new DependencyCacheIntegrityError(`retention-walk-failed:${errorCode(error) ?? "unknown"}`);
-      throw reportIntegrityRefusal(report, name, integrityError.condition);
-    }
-    const markerInfo = markerKinds.get(name) ?? null;
-    const usedMs = Number(markerInfo?.mtimeMs ?? info.mtimeMs);
-    if (!Number.isFinite(usedMs)) throw reportIntegrityRefusal(report, name, "invalid-usage-marker-time");
-    entries.push({ key: name, path, usedMs, bytes });
-  }
-  return entries;
-};
-
-const totalRetentionBytes = (entries: RetentionEntry[]): bigint =>
-  entries.reduce((total, entry) => total + entry.bytes, 0n);
-
-const removeRetentionEntry = async (
-  entry: RetentionEntry,
-  usageRoot: string,
-): Promise<void> => {
-  try {
-    await makeWritable(entry.path);
-    await rm(entry.path, { recursive: true, force: true });
-    if (await pathKind(entry.path) !== "missing") throw new Error("cache entry remained after deletion");
-    await rm(join(usageRoot, entry.key), { force: true });
-    if (await pathKind(join(usageRoot, entry.key)) !== "missing") {
-      throw new Error("cache usage marker remained after deletion");
-    }
-  } catch (error: unknown) {
-    throw new DependencyCacheBudgetError(`eviction-failed:${errorCode(error) ?? "unknown"}`);
-  }
-};
-
-const rollbackPublishedEntry = async (
-  cacheRoot: string,
-  key: string,
-  report: (progress: DependencyCacheProgress) => void,
-): Promise<void> => {
-  const entry: RetentionEntry = {
-    key,
-    path: join(cacheRoot, "entries", key),
-    usedMs: 0,
-    bytes: 0n,
-  };
-  const existed = await pathKind(entry.path) !== "missing";
-  try {
-    await removeRetentionEntry(entry, join(cacheRoot, CACHE_USAGE_DIRECTORY));
-  } catch (error: unknown) {
-    const condition = error instanceof DependencyCacheBudgetError
-      ? `publication-rollback-failed:${error.condition}`
-      : "publication-rollback-failed";
-    report({ event: "integrity-refusal", key: key.slice(0, 16), condition });
-    throw new DependencyCacheBudgetError(condition);
-  }
-  if (existed) report({ event: "eviction", key: key.slice(0, 16), condition: "byte-budget" });
-};
-
-const asRetentionIntegrityError = (error: unknown): DependencyCacheIntegrityError =>
-  error instanceof DependencyCacheIntegrityError
-    ? error
-    : new DependencyCacheIntegrityError(`retention-walk-failed:${errorCode(error) ?? "unknown"}`);
-
-export const enforceDependencyCacheByteBudget = async (
-  cacheRoot: string,
-  currentKey: string | undefined,
-  newlyPublishedKey: string | undefined,
-  report: (progress: DependencyCacheProgress) => void,
-  budget = DEPENDENCY_CACHE_BYTE_BUDGET,
-): Promise<void> => {
-  if (!Number.isSafeInteger(budget) || budget < 0) {
-    throw new DependencyCacheBudgetError("invalid-byte-budget");
-  }
-  const budgetBytes = BigInt(budget);
-  const handle = await openCacheLock(cacheRoot);
-  let locked = false;
-  try {
-    // This is deliberately blocking. Shared restore and publication owners
-    // must finish before this snapshot can be sized or entries removed.
-    await flockAsync(handle.fd, "ex");
-    locked = true;
-    try {
-      const usageRoot = join(cacheRoot, CACHE_USAGE_DIRECTORY);
-      let entries: RetentionEntry[];
-      try {
-        entries = await retentionEntries(cacheRoot, report);
-      } catch (error: unknown) {
-        const integrityError = asRetentionIntegrityError(error);
-        if (!(error instanceof DependencyCacheIntegrityError)) {
-          report({ event: "integrity-refusal", condition: integrityError.condition });
-        }
-        throw integrityError;
-      }
-      const protectedKey = currentKey && CACHE_KEY.test(currentKey) ? currentKey : undefined;
-      let total = totalRetentionBytes(entries);
-      if (total > budgetBytes) {
-        let victimKeys: string[];
-        try {
-          victimKeys = selectDependencyCacheEvictions(
-            entries.map(({ key, bytes, usedMs }) => {
-              const numericBytes = Number(bytes);
-              if (!Number.isSafeInteger(numericBytes)) throw new DependencyCacheIntegrityError("allocated-size-overflow");
-              return { key, bytes: numericBytes, usedMs };
-            }),
-            protectedKey,
-            budget,
-          );
-        } catch (error: unknown) {
-          if (error instanceof DependencyCacheBudgetError
-            && newlyPublishedKey
-            && CACHE_KEY.test(newlyPublishedKey)) {
-            await rollbackPublishedEntry(cacheRoot, newlyPublishedKey, report);
-          }
-          const condition = error instanceof DependencyCacheBudgetError
-            ? error.condition
-            : asRetentionIntegrityError(error).condition;
-          report({
-            event: "integrity-refusal",
-            ...(protectedKey ? { key: protectedKey.slice(0, 16) } : {}),
-            condition,
-          });
-          throw error instanceof DependencyCacheBudgetError ? error : new DependencyCacheIntegrityError(condition);
-        }
-        const entriesByKey = new Map(entries.map((entry) => [entry.key, entry] as const));
-        for (const victimKey of victimKeys) {
-          const victim = entriesByKey.get(victimKey);
-          if (victim === undefined) throw new DependencyCacheBudgetError("byte-budget-invariant-unmet");
-          try {
-            await removeRetentionEntry(victim, usageRoot);
-          } catch (error: unknown) {
-            const condition = error instanceof DependencyCacheBudgetError ? error.condition : "eviction-failed";
-            report({ event: "integrity-refusal", key: victim.key.slice(0, 16), condition });
-            if (newlyPublishedKey && newlyPublishedKey !== victim.key && CACHE_KEY.test(newlyPublishedKey)) {
-              await rollbackPublishedEntry(cacheRoot, newlyPublishedKey, report);
-            }
-            throw error instanceof DependencyCacheBudgetError ? error : new DependencyCacheBudgetError(condition);
-          }
-          total -= victim.bytes;
-          report({ event: "eviction", key: victim.key.slice(0, 16), condition: "byte-budget" });
-        }
-      }
-
-      // The exclusive lock has covered the initial walk and every deletion,
-      // so this tracked total is the authoritative final invariant without a
-      // second full manifest parse and inode walk.
-      if (total <= budgetBytes) return;
-
-      const protectedEntry = protectedKey === undefined
-        ? undefined
-        : entries.find((entry) => entry.key === protectedKey);
-      const condition = protectedEntry !== undefined && protectedEntry.bytes > budgetBytes
-        ? "protected-entry-exceeds-byte-budget"
-        : "byte-budget-invariant-unmet";
-      if (newlyPublishedKey && CACHE_KEY.test(newlyPublishedKey)) {
-        const published = entries.find((entry) => entry.key === newlyPublishedKey) ?? {
-          key: newlyPublishedKey,
-          path: join(cacheRoot, "entries", newlyPublishedKey),
-          usedMs: 0,
-          bytes: 0n,
-        } satisfies RetentionEntry;
-        await rollbackPublishedEntry(cacheRoot, published.key, report);
-      }
-      report({
-        event: "integrity-refusal",
-        ...(protectedKey ? { key: protectedKey.slice(0, 16) } : {}),
-        condition,
-      });
-      throw new DependencyCacheBudgetError(condition);
-    } finally {
-      await flockAsync(handle.fd, "un");
-      locked = false;
-    }
-  } finally {
-    if (locked) await flockAsync(handle.fd, "un").catch(() => undefined);
-    await handle.close();
-  }
-};
-
 const restoreEntry = async (
   config: RunnerConfig,
-  metadata: CacheMetadata,
-  entry: string,
+  store: CacheEntryStore,
+  document: CacheEntryDocument,
   workspace: string,
   env: NodeJS.ProcessEnv,
   execute: DependencyCommandExecutor,
 ): Promise<void> => {
-  await clearTargets(config, workspace, metadata.targets.map(({ path }) => path), env, execute);
+  await clearTargets(config, workspace, document.targets.map(({ path }) => path), env, execute);
   try {
-    for (const target of metadata.targets) {
+    for (const target of document.targets) {
       if (!target.present) continue;
-      const source = resolve(entry, TREE_DIRECTORY, target.path);
-      const destination = assertTarget(workspace, target.path);
+      const source = store.targetSourcePath(document.key, target.path);
+      const destination = assertCacheTargetPath(workspace, target.path);
       const args = platform() === "darwin"
         ? ["-c", "-R", source, destination]
         : ["-a", "--reflink=always", source, destination];
       await execute(config, "/bin/cp", args, workspace, env);
       await execute(config, "/bin/chmod", ["-R", "u+w", destination], workspace, env);
     }
-    const restored = await findTargetManifest(workspace, metadata.targets.map(({ path }) => path));
-    if (!sameJson(restored, metadata.targets)) throw new Error("Restored dependency target manifest differs from the cache entry");
+    const restored = await describeTargetTrees(workspace, document.targets.map(({ path }) => path));
+    if (!sameJson(restored, document.targets)) throw new Error("Restored dependency target manifest differs from the cache entry");
   } catch (error: unknown) {
-    await clearTargets(config, workspace, metadata.targets.map(({ path }) => path), env, execute).catch(() => undefined);
+    await clearTargets(config, workspace, document.targets.map(({ path }) => path), env, execute).catch(() => undefined);
     throw error;
-  }
-};
-
-const snapshotTarget = async (source: string, destination: string): Promise<void> => {
-  await mkdir(dirname(destination), { recursive: true });
-  await cp(source, destination, {
-    recursive: true,
-    preserveTimestamps: true,
-    verbatimSymlinks: true,
-    mode: constants.COPYFILE_FICLONE,
-  });
-};
-
-const publishEntry = async (
-  cacheRoot: string,
-  entry: string,
-  metadata: CacheMetadata,
-  workspace: string,
-  expected: Omit<CacheMetadata, "targets"> & { targetPaths: string[] },
-): Promise<"published" | "converged" | "refused"> => {
-  if (await pathKind(entry) !== "missing") {
-    try {
-      await validateEntry(entry, expected, workspace);
-      return "converged";
-    } catch (error: unknown) {
-      if (error instanceof DependencyCacheIntegrityError) return "refused";
-      throw error;
-    }
-  }
-  const stagingParent = dirname(entry);
-  const staging = await mkdtemp(join(stagingParent, `.stage-${metadata.key.slice(0, 16)}-`));
-  if (!insideOrEqual(cacheRoot, await realpath(staging))) throw new Error("Dependency cache staging escaped its root");
-  try {
-    await mkdir(join(staging, TREE_DIRECTORY));
-    for (const target of metadata.targets) {
-      if (target.present) await snapshotTarget(assertTarget(workspace, target.path), resolve(staging, TREE_DIRECTORY, target.path));
-    }
-    await writeFile(join(staging, METADATA_FILE), `${JSON.stringify(metadata)}\n`, { mode: 0o400 });
-    await makeImmutable(staging);
-    await validateEntry(staging, expected, workspace);
-    try {
-      await rename(staging, entry);
-      return "published";
-    } catch (error: unknown) {
-      // Darwin reports EACCES rather than EEXIST when the winning directory is
-      // already immutable. The destination's independently verified state,
-      // not the platform-specific errno, decides whether this was a safe race.
-      if (await pathKind(entry) === "missing") throw error;
-      try {
-        await validateEntry(entry, expected, workspace);
-        return "converged";
-      } catch (validationError: unknown) {
-        if (validationError instanceof DependencyCacheIntegrityError) return "refused";
-        throw validationError;
-      }
-    }
-  } finally {
-    if (await pathKind(staging) !== "missing") {
-      await makeWritable(staging).catch(() => undefined);
-      await rm(staging, { recursive: true, force: true });
-    }
   }
 };
 
@@ -1284,7 +558,7 @@ const installDependencies = async (
   env: NodeJS.ProcessEnv,
   execute: DependencyCommandExecutor,
   retryOptions: RetryOptions = {},
-): Promise<TargetManifestEntry[]> => {
+): Promise<CacheEntryTarget[]> => {
   const args = ["ci", "--prefer-offline", "--no-audit", "--no-fund"];
   try {
     await runWithNetworkRetry("npm", args, async ({ timeoutMs }) => {
@@ -1295,7 +569,7 @@ const installDependencies = async (
       budgetMs: NPM_INSTALL_OPERATION_BUDGET_MS,
       ...retryOptions,
     });
-    return await findTargetManifest(workspace, targets);
+    return await installedTargetTrees(workspace, targets);
   } catch (error: unknown) {
     await clearTargets(config, workspace, targets, env, execute).catch(() => undefined);
     throw error;
@@ -1360,27 +634,25 @@ export const materializeWorkspaceDependencies = async (
       throw error;
     }
     const { key, toolchain, inputs, targets: targetPaths, nativeWorkspaces } = identity;
-    const cacheRoot = await ensureCacheRoot(
+    const store = await openCacheEntryStore(
       options.cacheRoot ?? config.dependencyCacheRoot ?? join(dirname(resolve(config.workspaceRoot)), "dependency-cache"),
       workspace,
     );
-    const entry = resolve(cacheRoot, "entries", key);
-    if (!insideOrEqual(cacheRoot, entry)) throw new Error("Dependency cache entry escaped its root");
-    const expected = { format: CACHE_FORMAT, key, toolchain, inputs, targetPaths } as const;
-    const locked = await withSharedCacheLock(cacheRoot, async () => {
+    const expected: CacheEntryExpectation = { key, toolchain, inputs, targetPaths };
+    const locked = await store.withSharedLock(async () => {
       try {
-        await validateCacheUseMarker(cacheRoot, key);
+        await store.validateUseMarker(key);
       } catch (error: unknown) {
         if (!(error instanceof DependencyCacheIntegrityError)) throw error;
         report({ event: "integrity-refusal", key: key.slice(0, 16), condition: error.condition });
         throw error;
       }
-      if (await pathKind(entry) !== "missing") {
+      if (await store.hasEntry(key)) {
         try {
-          const metadata = await timed("validate", report, () => validateEntry(entry, expected, workspace));
-          await timed("restore", report, () => restoreEntry(config, metadata, entry, workspace, env, execute));
+          const document = await timed("validate", report, () => store.readEntry(expected));
+          await timed("restore", report, () => restoreEntry(config, store, document, workspace, env, execute));
           await timed("rebuild", report, () => rebuildNativeWorkspaces(config, workspace, nativeWorkspaces, env, execute));
-          await recordCacheUse(cacheRoot, key);
+          await store.recordUse(key);
           return {
             result: { status: "restored", key } as DependencyCacheResult,
             usableKey: key,
@@ -1399,15 +671,14 @@ export const materializeWorkspaceDependencies = async (
       const targets = await timed("install", report, () => installDependencies(
         config, workspace, targetPaths, env, execute, options.installRetryOptions,
       ));
-      const metadata: CacheMetadata = { format: CACHE_FORMAT, key, toolchain, inputs, targets };
-      const publication = await timed("publish", report, () => publishEntry(cacheRoot, entry, metadata, workspace, expected));
+      const publication = await timed("publish", report, () => store.publishEntry(expected, targets, workspace));
       if (publication === "refused") {
         const integrityError = new DependencyCacheIntegrityError("concurrent-entry-invalid");
         report({ event: "integrity-refusal", key: key.slice(0, 16), condition: integrityError.condition });
         throw integrityError;
       }
       try {
-        await recordCacheUse(cacheRoot, key);
+        await store.recordUse(key);
       } catch (error: unknown) {
         if (!(error instanceof DependencyCacheIntegrityError)) throw error;
         report({ event: "integrity-refusal", key: key.slice(0, 16), condition: error.condition });
@@ -1422,8 +693,7 @@ export const materializeWorkspaceDependencies = async (
         } as DependencyCacheProgress,
       };
     });
-    await timed("retention", report, () => enforceDependencyCacheByteBudget(
-      cacheRoot,
+    await timed("retention", report, () => store.enforceByteBudget(
       locked.usableKey,
       locked.newlyPublishedKey,
       report,
